Rename plugin to Multicombat borders.
