+++ runelite-client/src/main/java/net/runelite/client/plugins/multilines/MultiCombatBordersConfig.java
@@ -24,41 +24,37 @@
  */
 package net.runelite.client.plugins.multilines;
 
-import lombok.Value;
-import net.runelite.api.coords.WorldPoint;
+import java.awt.Color;
+import net.runelite.client.config.Config;
+import net.runelite.client.config.ConfigGroup;
+import net.runelite.client.config.ConfigItem;
 
-@Value
-public class MultiLineBorder
+@ConfigGroup(
+	name = "Multicombat borders",
+	keyName = "multiCombatBorders",
+	description = "Configuration for the multicombat borders plugin"
+)
+public interface MultiCombatBordersConfig extends Config
 {
 
-	public static final int NORTH = 0;
-	public static final int SOUTH = 1;
-	public static final int EAST = 2;
-	public static final int WEST = 3;
-
-	private final String id;
-	private final WorldPoint single;
-	private final WorldPoint multi;
-
-	int getEdge()
+	@ConfigItem(
+		keyName = "singleColor",
+		name = "Single color",
+		description = "Color of singlecombat side of border"
+	)
+	default Color singleColor()
 	{
-		if (single.getX() > multi.getX())
-		{
-			return EAST;
-		}
-		else if (single.getX() < multi.getX())
-		{
-			return WEST;
-		}
-		else if (single.getY() > multi.getY())
-		{
-			return NORTH;
-		}
-		else if (single.getY() < multi.getY())
-		{
-			return SOUTH;
-		}
-		throw new IllegalStateException("Entry is not representing a border.");
+		return Color.GREEN;
+	}
+
+	@ConfigItem(
+		keyName = "multiColor",
+		name = "Multi color",
+		description = "Color of multicombat side of border"
+	)
+	default Color multiColor()
+	{
+		return Color.RED;
 	}
 
 }
